Added two options to catenate for specifying a listing and omitting deletion upon request.

+++ postmutect/catenate.py
@@ -16,6 +16,14 @@ except ImportError as I:
 
 parser = makeparser(('Concatenates the vcf files within the subdirectories'
                      ' created by multimutect.'))
+parser.add_argument('--delete_fragments', action='store_true',
+                    help='Delete files utilized for catenation')
+parser.add_argument('-l', '--listing', type=str,
+                    help=('A path specifying the list of samples to'
+                          'concatenate. Useful if the BAM files used'
+                          'for the analysis were created on a per-'
+                          'chromosome basis. Uses chrs.list by default.')
+                    default='chrs.list')
 args = parser.parse_args()
 
 """
@@ -33,7 +41,7 @@ At the moment, cleans up after multimutect.py with regards to the
 creation of a status directory as well as moving status files up into
 said directory.
 """
-def vcf_catenate(parent, reference, gatkpath='gatk.jar'):
+def vcf_catenate(parent, reference, gatkpath, listfile, delete):
     cmd = ('java -cp {gatk} org.broadinstitute.gatk.tools.CatVariants'
            ' -assumeSorted -R' 
            ' {ref} -out {{out}} {{vcfs}}').format(gatk=gatkpath, 
@@ -41,12 +49,9 @@ def vcf_catenate(parent, reference, gatkpath='gatk.jar'):
     for dirpath, dirnames, filenames in os.walk(parent):
         for directory in dirnames:
             d_path = os.path.join(dirpath, directory)
-            listing = os.path.join(d_path, 'chrs.list')
-            #NOTE: An underscore is no longer at the end of the 
-            #directories. However I am going to leave this here just in
-            #case.
-            result_name = directory.strip('_') + '.vcf'
-            status_name = directory.strip('_') + '.list'
+            listing = os.path.join(d_path, listfile)
+            result_name = directory + '.vcf'
+            status_name = directory + '.list'
             outpath = os.path.join(dirpath, result_name)
             #Make sure the listing file exists and that this isn't the
             #status directory.
@@ -61,18 +66,20 @@ def vcf_catenate(parent, reference, gatkpath='gatk.jar'):
                     vseries = "".join(['-V ' + c + ' ' for c in realchrs])
                     final_cmd = cmd.format(out=outpath, vcfs=vseries)
                     check_output(final_cmd.split())
-                    #Clean up leftover files after combining them.
+                    #Clean up leftover files after combining them if the option
+                    #is specified.
                     #Also cleans up .idx files and directories.
-                    map(os.unlink, chrlist)
-                    #No index file is created for empty vcfs.
-                    map(lambda x: os.unlink(x + '.idx'), realchrs)
-                    os.unlink(listing)
-            os.rmdir(d_path)
+                    if delete == True:
+                        map(os.unlink, chrlist)
+                        #No index file is created for empty vcfs.
+                        map(lambda x: os.unlink(x + '.idx'), realchrs)
+                        os.unlink(listing)
+                        os.rmdir(d_path)
 
 if os.path.exists(args.gatkpath):
     vcf_catenate(args.directory, args.reference, args.gatkpath)
-elif os.exists('gatk.jar'):
-    vcf_catenate(args.directory, args.reference)
+elif os.path.exists('gatk.jar'):
+    vcf_catenate(args.directory, args.reference, args.gatkpath)
 else:
     sys.stderr.write('Please provide an existent gatk jar filepath.')
     sys.exit(1)
